requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0
scipy>=1.12.0

# Delta Lake (local, no Spark needed)
deltalake>=0.17.0

# ML
scikit-learn>=1.4.0
mlflow>=2.12.0
shap>=0.45.0

# AI / Genie QA
anthropic>=0.25.0

# UI
streamlit>=1.34.0
plotly>=5.20.0

# SQL execution for Genie QA (reads Delta Parquet files)
duckdb>=0.10.0

# Config
python-dotenv>=1.0.0

# Optional: Databricks connectivity
# databricks-sdk>=0.20.0
